refactor mapSection, 안개 관련 코드 삭제
