+++ app/src/main/java/com/dudoji/android/map/domain/MapSection.kt
@@ -12,7 +12,7 @@ import androidx.core.graphics.get
 import com.dudoji.android.config.BASIC_ZOOM_LEVEL
 import com.dudoji.android.config.FOG_COLOR
 import com.dudoji.android.config.GRADIENT_RADIUS_RATE
-import com.dudoji.android.map.utils.tile.TILE_SIZE
+import com.dudoji.android.config.TILE_SIZE
 import com.dudoji.android.map.utils.tile.TileCoordinateUtil
 
 open class MapSection {
