+++ app/src/main/java/com/dudoji/android/map/repository/RangeSearchDataSource.kt
@@ -1,10 +1,10 @@
 package com.dudoji.android.map.repository
 
+import android.location.Location
 import android.os.Build
 import android.util.Log
 import androidx.annotation.RequiresApi
 import com.dudoji.android.config.PIN_UPDATE_THRESHOLD
-import com.dudoji.android.map.utils.MapUtil
 import com.dudoji.android.network.api.dto.BaseDto
 import com.google.android.gms.maps.model.LatLng
 import retrofit2.Response
@@ -20,10 +20,16 @@ abstract class RangeSearchDataSource<T:BaseDto<D>, D> {
         radius: Double
     ): Response<List<T>>
 
+    fun distanceBetween(p1: LatLng, p2: LatLng): Double {
+        val results = FloatArray(1)
+        Location.distanceBetween(p1.latitude, p1.longitude, p2.latitude, p2.longitude, results)
+        return results[0].toDouble()
+    }
+    
     @RequiresApi(Build.VERSION_CODES.O)
     suspend fun load(latLng: LatLng, radius: Double): Boolean{
         if (lastUpdatedLatLng == null ||
-            MapUtil.distanceBetween(lastUpdatedLatLng!!, latLng) > PIN_UPDATE_THRESHOLD) {
+            distanceBetween(lastUpdatedLatLng!!, latLng) > PIN_UPDATE_THRESHOLD) {
             val response = fetchFromApi(
                 latLng.latitude,
                 latLng.longitude,
