+++ app/src/main/java/com/dudoji/android/map/utils/npc/NpcApplier.kt
@@ -4,11 +4,11 @@ import android.graphics.BitmapFactory
 import android.os.Build
 import androidx.annotation.RequiresApi
 import com.dudoji.android.R
-import com.dudoji.android.presentation.map.MapActivity
-import com.dudoji.android.presentation.map.MapActivity.ActivityMapObject
+import com.dudoji.android.domain.model.ActivityMapObject
 import com.dudoji.android.map.domain.Npc
 import com.dudoji.android.map.repository.NpcDataSource
 import com.dudoji.android.map.utils.NonClusterMarkerApplier
+import com.dudoji.android.presentation.map.MapActivity
 import com.google.android.gms.maps.GoogleMap
 import com.google.android.gms.maps.GoogleMap.OnCameraIdleListener
 import com.google.android.gms.maps.model.LatLng
@@ -40,7 +40,8 @@ class NpcApplier(normalMarkerCollection: MarkerManager.Collection, googleMap: Go
         val npc = marker?.tag as? Npc ?: return
         if (npc.hasQuest) {
             val initialBitmap = BitmapFactory.decodeResource(activity.resources, R.mipmap.quest_bubble)
-            val activityMapObject = ActivityMapObject(LatLng(npc.lat, npc.lng), initialBitmap, 70f, -150f, 170, -100)
+            val activityMapObject =
+                ActivityMapObject(LatLng(npc.lat, npc.lng), initialBitmap, 70f, -150f, 170, -100)
             npc.activityMapObject = activityMapObject
             activity.activityObjects.add(activityMapObject)
         }
