--- app/src/main/java/com/dudoji/android/map/utils/tile/MaskTileProvider.kt
@@ -1,22 +1,0 @@
-package com.dudoji.android.map.utils.tile
-
-
-import android.graphics.Bitmap
-import com.dudoji.android.map.utils.tile.mask.IMaskTileMaker
-import com.google.android.gms.maps.model.*
-import kotlinx.coroutines.runBlocking
-import java.io.ByteArrayOutputStream
-
-const val TILE_SIZE = 512
-
-// Mask Tile Provider for Map Activity
-class MaskTileProvider(private val maskTileMaker: IMaskTileMaker) : TileProvider {
-
-    override fun getTile(x : Int, y : Int, zoom : Int) : Tile {
-        val bitmap : Bitmap = runBlocking { maskTileMaker.createMaskTile(x, y, zoom) }
-        val stream = ByteArrayOutputStream()
-        bitmap.compress(Bitmap.CompressFormat.PNG, 100, stream)
-        val byteArray = stream.toByteArray()
-        return Tile(TILE_SIZE, TILE_SIZE, byteArray)
-    }
-}
