+++ app/src/main/java/com/dudoji/android/pin/util/PinFilter.kt
@@ -6,14 +6,12 @@ import androidx.annotation.RequiresApi
 import androidx.appcompat.app.AppCompatActivity
 import coil.load
 import com.dudoji.android.R
-import com.dudoji.android.presentation.map.MapActivity
-import com.dudoji.android.map.manager.DatabaseMapSectionManager
 import com.dudoji.android.pin.domain.Pin
 import com.dudoji.android.pin.domain.Who
+import com.dudoji.android.presentation.map.MapActivity
 
 class PinFilter(
     private val activity: AppCompatActivity,
-    private val databaseMapSectionManager: DatabaseMapSectionManager?
 ) {
     private lateinit var btnMine: ImageButton
     private lateinit var btnFriend: ImageButton
@@ -28,11 +26,10 @@ class PinFilter(
     fun filterPins(pins: List<Pin>): List<Pin> {
         return pins.filter { pin ->
             visibilityMap[pin.master] == true
-                    && !(databaseMapSectionManager?.isFogExists(pin.lat, pin.lng) ?: false)
         }
     }
 
-    @RequiresApi(Build.VERSION_CODES.O)
+
     fun setupFilterButtons() {
         btnMine = activity.findViewById(R.id.btnFilterMine)
         btnFriend = activity.findViewById(R.id.btnFilterFriend)
